requirements.txt.txt
# AI Hardware Debug Assistant - Dependencies
# Install with: pip install -r requirements.txt

streamlit>=1.28.0
requests>=2.31.0
Pillow>=10.0.0
